requirements.txt
pyautogui>=0.9.54
pycaw>=20240210
# pywin32>=308
# pywin32-ctypes>=0.2.2
pywinauto>=0.6.8
psutil>=7.0.0

SQLAlchemy[asyncio]==2.0.31
pydantic-settings==2.6.1
pydantic-validation-decorator==0.1.4
APScheduler==3.10.4
DateTime==5.5
openpyxl==3.1.5
passlib[bcrypt]==1.7.4
tinify==1.6.0
pandas==2.0.3
PyYaml==6.0.1
python-dotenv>=1.1.0
motor>=3.3.0

pywebview>=4.0.0
cozepy==0.20.0
aiosqlite>=0.17.0
